requirements.txt
httpx
beautifulsoup4
google-cloud-bigquery
google-auth
langdetect
tqdm
openai
python-dotenv
streamlit
plotly
pandas
db-dtypes